extract hot key trigger to its own hook
