+++ src/Tool.js
@@ -1,7 +1,8 @@
-import React, { useCallback, useEffect } from "react";
+import React, { useCallback } from "react";
 import { useGlobals } from "@storybook/api";
 import { Icons, IconButton } from "@storybook/components";
 import { TOOL_ID } from "./constants";
+import { useHotKey } from "./useHotKey";
 
 export const Tool = () => {
   const [{ measureEnabled }, updateGlobals] = useGlobals();
@@ -14,25 +15,7 @@ export const Tool = () => {
     [measureEnabled]
   );
 
-  useEffect(() => {
-    const onKeyDown = (e) => {
-      if (e.key === "Alt") {
-        updateGlobals({ measureEnabled: true });
-      }
-    };
-
-    const onKeyUp = () => {
-      updateGlobals({ measureEnabled: false });
-    };
-
-    document.addEventListener("keydown", onKeyDown);
-    document.addEventListener("keyup", onKeyUp);
-
-    return () => {
-      document.removeEventListener("keydown", onKeyDown);
-      document.removeEventListener("keyup", onKeyUp);
-    };
-  }, []);
+  useHotKey(updateGlobals);
 
   return (
     <IconButton
